+++ hc/api/urls.py
@@ -25,6 +25,7 @@ urlpatterns = [
     path("api/v1/checks/<uuid:code>", views.single, name="hc-api-single"),
     path("api/v1/checks/<uuid:code>/pause", views.pause, name="hc-api-pause"),
     path("api/v1/notifications/<uuid:code>/bounce", views.bounce, name="hc-api-bounce"),
+    path("api/v1/checks/<uuid:code>/pings/", views.pings, name="hc-api-pings"),
     path("api/v1/channels/", views.channels),
     path(
         "badge/<slug:badge_key>/<slug:signature>/<quoted:tag>.<slug:fmt>",