+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Paused ping handling can be controlled via API (#376)
+- Add "Get a list of checks's logged pings" API call (#371)
+
 
 ### Bug Fixes
 - Removing Pager Team integration, project appears to be discontinued